pyproject.toml
[tool.poetry]
name = "vga-verilog"
version = "0.1.0"
description = ""
authors = ["Soham Jaiswal <sohamjaiswal10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
